backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
celery[redis]
requests
aiohttp
dnspython
passlib[bcrypt]
python-jose[cryptography]
httpx
pytest
pytest-asyncio
pytest-cov
email-validator
argon2-cffi
flower
python-dotenv